refactor: relocate dto files
