--- src/main/kotlin/ecommerce/dto/product/ProductPatchRequest.kt
@@ -1,7 +1,0 @@
-package ecommerce.dto.product
-
-data class ProductPatchRequest(
-    val name: String? = null,
-    val price: Double? = null,
-    val imageUrl: String? = null,
-)
